+++ packages/core/cypress/integration/plp.test.js
@@ -115,7 +115,7 @@ describe('Infinite Scroll pagination', () => {
     cy.waitForHydration()
 
     cy.getById('product-gallery').within(() => {
-      cy.getById('store-card')
+      cy.getById('store-product-card')
         .should('exist')
         .should('have.length.gt', 0)
         .then(($links) => {
@@ -125,7 +125,7 @@ describe('Infinite Scroll pagination', () => {
             .should('exist')
             .click()
             .then(() => {
-              cy.getById('store-card')
+              cy.getById('store-product-card')
                 .should('have.length.gte', before)
                 .then(($products) => {
                   const after = $products.length
@@ -141,7 +141,7 @@ describe('Infinite Scroll pagination', () => {
     cy.visit(pages.collection, options)
     cy.waitForHydration()
 
-    cy.get('[data-testid=product-gallery] [data-testid=store-card]')
+    cy.get('[data-testid=product-gallery] [data-testid=store-product-card]')
       .should('exist')
       .should('have.length.gt', 0)
       .then(($links) => {
@@ -152,12 +152,12 @@ describe('Infinite Scroll pagination', () => {
           .should('exist')
           .click()
           // Go up the page
-          .get('[data-testid=product-gallery] [data-testid=store-card]')
+          .get('[data-testid=product-gallery] [data-testid=store-product-card]')
           .first()
           .scrollIntoView({ duration: 1000 })
 
           // Go down the page
-          .get('[data-testid=product-gallery] [data-testid=store-card]')
+          .get('[data-testid=product-gallery] [data-testid=store-product-card]')
           .last()
           .scrollIntoView({ duration: 1000 })
           .then(() => {
@@ -167,7 +167,9 @@ describe('Infinite Scroll pagination', () => {
             // The skuId of the last product on the page
             let skuIdBeforeNavigate
 
-            cy.get('[data-testid=product-gallery] [data-testid=store-card]')
+            cy.get(
+              '[data-testid=product-gallery] [data-testid=store-product-card]'
+            )
               // Number of products after showMore is clicked should be higher
               .should('have.length.gte', before)
               .last()
@@ -183,7 +185,9 @@ describe('Infinite Scroll pagination', () => {
               })
               .then(() => {
                 cy.go('back')
-                  .get('[data-testid=product-gallery] [data-testid=store-card]')
+                  .get(
+                    '[data-testid=product-gallery] [data-testid=store-product-card]'
+                  )
                   .last()
                   .then(($card) => {
                     const skuIdAfterNavigate = $card.attr(
@@ -212,7 +216,7 @@ describe('Infinite Scroll pagination', () => {
       .location('search')
       .should('match', /page=0$/)
 
-      .get('[data-testid=product-gallery] [data-testid=store-card]')
+      .get('[data-testid=product-gallery] [data-testid=store-product-card]')
       .last()
       .scrollIntoView({ duration: 1000 })
       .location('search')